+++ datax-web/src/main/java/com/wugui/tool/meta/DatabaseInterface.java
@@ -3,7 +3,7 @@ package com.wugui.tool.meta;
 public interface DatabaseInterface {
 
     /**
-     * Returns the minimal SQL to launch in order to determine the layout of the resultset for a given database table
+     * Returns the minimal SQL to launch in order to determine the layout of the resultset for a given com.wugui.database table
      *
      * @param tableName The name of the table to determine the layout for
      * @return The SQL to launch.
