+++ datax-web/src/main/java/com/wugui/tool/query/QueryToolInterface.java
@@ -1,7 +1,7 @@
 package com.wugui.tool.query;
 
-import database.ColumnInfo;
-import database.TableInfo;
+import com.wugui.database.ColumnInfo;
+import com.wugui.database.TableInfo;
 
 import java.util.List;
 import java.util.Map;
