+++ datax-web/src/main/java/com/wugui/tool/util/DasUtil.java
@@ -3,7 +3,7 @@ package com.wugui.tool.util;
 import com.wugui.dataxweb.entity.JobJdbcDatasource;
 import com.wugui.tool.query.BaseQueryTool;
 import com.wugui.tool.query.QueryToolFactory;
-import database.TableInfo;
+import com.wugui.database.TableInfo;
 import lombok.extern.slf4j.Slf4j;
 
 /**
